figures dir added

+++ app.py
@@ -42,6 +42,7 @@ CREATE TABLE IF NOT EXISTS stock_prices (date DATE, open DOUBLE, high DOUBLE, lo
 CREATE TABLE IF NOT EXISTS income_statement( date DATE, symbol VARCHAR(20), reportedCurrency VARCHAR(20), cik VARCHAR(20), fillingDate DATE, acceptedDate DATE, calendarYear INTEGER, period VARCHAR(20), revenue INTEGER, costOfRevenue INTEGER, grossProfit INTEGER, grossProfitRatio DOUBLE, researchAndDevelopmentExpenses INTEGER, generalAndAdministrativeExpenses INTEGER, sellingAndMarketingExpenses INTEGER, sellingGeneralAndAdministrativeExpenses INTEGER, otherExpenses INTEGER, operatingExpenses INTEGER, costAndExpenses INTEGER, interestIncome INTEGER, interestExpense INTEGER, depreciationAndAmortization INTEGER, ebitda INTEGER, ebitdaratio DOUBLE, operatingIncome INTEGER, operatingIncomeRatio DOUBLE, totalOtherIncomeExpensesNet INTEGER, incomeBeforeTax INTEGER, incomeBeforeTaxRatio DOUBLE, incomeTaxExpense INTEGER, netIncome INTEGER, netIncomeRatio DOUBLE, eps DOUBLE, epsdiluted DOUBLE, weightedAverageShsOut INTEGER, weightedAverageShsOutDil INTEGER );
 CREATE TABLE IF NOT EXISTS balancesheet_statement( date DATE, symbol VARCHAR(20), reportedCurrency VARCHAR(20), cik VARCHAR(20), fillingDate DATE, acceptedDate DATE, calendarYear INTEGER, period VARCHAR(20), cashAndCashEquivalents INTEGER, shortTermInvestments INTEGER, cashAndShortTermInvestments INTEGER, netReceivables INTEGER, inventory INTEGER, otherCurrentAssets INTEGER, totalCurrentAssets INTEGER, propertyPlantEquipmentNet INTEGER, goodwill INTEGER, intangibleAssets INTEGER, goodwillAndIntangibleAssets INTEGER, longTermInvestments INTEGER, taxAssets INTEGER, otherNonCurrentAssets INTEGER, totalNonCurrentAssets INTEGER, otherAssets INTEGER, totalAssets INTEGER, accountPayables INTEGER, shortTermDebt INTEGER, taxPayables INTEGER, deferredRevenue INTEGER, otherCurrentLiabilities INTEGER, totalCurrentLiabilities INTEGER, longTermDebt INTEGER, deferredRevenueNonCurrent INTEGER, deferredTaxLiabilitiesNonCurrent INTEGER, otherNonCurrentLiabilities INTEGER, totalNonCurrentLiabilities INTEGER, otherLiabilities INTEGER, capitalLeaseObligations INTEGER, totalLiabilities VARCHAR(20), preferredStock VARCHAR(20), commonStock VARCHAR(20), retainedEarnings VARCHAR(20), accumulatedOtherComprehensiveIncomeLoss VARCHAR(20), othertotalStockholdersEquity VARCHAR(20), totalStockholdersEquity VARCHAR(20), totalEquity VARCHAR(20), totalLiabilitiesAndStockholdersEquity VARCHAR(20), minorityInterest VARCHAR(20), totalLiabilitiesAndTotalEquity VARCHAR(20), totalInvestments VARCHAR(20), totalDebt VARCHAR(20), netDebt VARCHAR(20) ); 
 CREATE TABLE IF NOT EXISTS cashflow_statement( date DATE, symbol VARCHAR(20), reportedCurrency VARCHAR(20), cik VARCHAR(20), fillingDate DATE, acceptedDate DATE, calendarYear INTEGER, period VARCHAR(20), netIncome INTEGER, depreciationAndAmortization INTEGER, deferredIncomeTax INTEGER, stockBasedCompensation INTEGER, changeInWorkingCapital INTEGER, accountsReceivables INTEGER, inventory INTEGER, accountsPayables INTEGER, otherWorkingCapital INTEGER, otherNonCashItems INTEGER, netCashProvidedByOperatingActivities INTEGER, investmentsInPropertyPlantAndEquipment INTEGER, acquisitionsNet INTEGER, purchasesOfInvestments INTEGER, salesMaturitiesOfInvestments INTEGER, otherInvestingActivites INTEGER, netCashUsedForInvestingActivites INTEGER, debtRepayment INTEGER, commonStockIssued INTEGER, commonStockRepurchased INTEGER, dividendsPaid INTEGER, otherFinancingActivites INTEGER, netCashUsedProvidedByFinancingActivities INTEGER, effectOfForexChangesOnCash INTEGER, netChangeInCash INTEGER, cashAtEndOfPeriod INTEGER, cashAtBeginningOfPeriod INTEGER, operatingCashFlow INTEGER, capitalExpenditure INTEGER, freeCashFlow INTEGER ); 
+The 'symbol' column in each table contains the companies names in capital letters.
 
 Request: {request}
 SQLQuery:
@@ -75,6 +76,7 @@ repl_tool.run("1+1")
 template2 = """
 Use the following pieces of user request and sql query to generate python code that should first load the required data from 'stock_db.sqlite' database and 
 then show insights related to that data. If the generated insights contains a figure or plot then that should be saved inside the 'figures' directory.
+If there is some tables or numerical values as insights then those should be printed out explicitely using print statement along with their description.
 Generate and return python code only, no additional text.
 If you don't know the answer, just say that you don't know, don't try to make up an answer.
 
@@ -91,49 +93,44 @@ chain2 = (PROMPT2
           | StrOutputParser()       # to get output in a more usable format
           )
 
+# Final Chain
 final_chain = chain1 | chain2
 
 
 # Create UI using Chainlit
 import chainlit as cl
 
-@cl.on_start
-def start():
-    # Set the title of the Chainlit app
-    cl.Title("My Application")
-
-@cl.on_chat_start
-async def start_chat():
-    await cl.Message("Welcome! Enter the query for stock prices.")
-
 @cl.on_message
 async def main(message: cl.Message):
+
+    # print(f"User Input: {message.content}")
 
     # Remove any files from 'figures' directory
     for filename in os.listdir('figures'):
         file_path = os.path.join('figures', filename)
-        if os.path.isfile(file_path):
+        if os.path.isfile(file_path) and filename != '__init__.py':
             os.remove(file_path)  # Remove file
 
-    print(f"Message: {message}")
-    response = final_chain.invoke({"request": message})
-    print(response)
-    print("="*50)
-    output = repl_tool.run(response)
-    print(output)
-
-    # Send response
-    # await cl.Message(content=f"Received: {message.content}",).send()
+    # Generate code for insights
+    code_response = final_chain.invoke({"request": message.content})
+    # # print(code_response)
+    # # print("="*50)
+    
+    # Execute code
+    output = repl_tool.run(code_response)
+    #print(output)
 
     # Send the plot to the chat
-    for imgfile in os.listdir('figures'): 
-        if os.path.isfile(os.path.join('figures', imgfile)):
-            # Attach the image to the message
-            image = cl.Image(path="./figures/"+imgfile, name="image1", display="inline")
-            await cl.Message(
-                content=f"Response: {output}",
-                elements=[image],
-            ).send()
-        else:
-            # Send response
-            await cl.Message(content=f"Response: {output}",).send()
+    if len(os.listdir('figures')) != 0:
+        for imgfile in os.listdir('figures'): 
+            print(imgfile)
+            if os.path.isfile(os.path.join('figures', imgfile)) and imgfile != '__init__.py':
+                # Attach the image to the message
+                image = cl.Image(path="./figures/"+imgfile, name="image1", size="large", display="inline")
+                await cl.Message(
+                    content=f"Response: \n{output}",
+                    elements=[image],
+                ).send()
+    else:
+        # Send response
+        await cl.Message(content=f"Response: \n{output}",).send()
